+++ pkg/envparser/parser_test.go
@@ -1,86 +1,8 @@
 package envparser
 
 import (
-	"os"
-	"path/filepath"
-	"strings"
 	"testing"
 )
-
-func TestGetEnvFiles(t *testing.T) {
-	tempDir, err := os.MkdirTemp("", "envTest")
-	if err != nil {
-		t.Fatalf("Failed to create temp dir: %v", err)
-	}
-	envDir := filepath.Join(tempDir, "env")
-	err = os.Mkdir(envDir, 0755)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	oldDir, err := os.Getwd()
-	if err != nil {
-		t.Fatalf("Failed to get current working directory: %v", err)
-	}
-	// Cleanup temp dir
-	defer func() {
-		err := os.Chdir(oldDir)
-		if err != nil {
-			t.Fatal(err)
-		}
-		os.RemoveAll(tempDir)
-	}()
-
-	// Test Cases
-	testFileNames := []string{"global.env", "test.ENV", "spec_pm.env"}
-
-	for _, fName := range testFileNames {
-		file, err := os.Create(filepath.Join(envDir, fName))
-		if err != nil {
-			t.Fatalf("Failed to Createfile :%v", err)
-		}
-		file.Close()
-	}
-
-	// Test Case: Directories should not be included:
-	err = os.Mkdir(filepath.Join(envDir, "shouldNotAppear"), 0755)
-	if err != nil {
-		t.Fatalf("Temporary directory could not be created: %v", err)
-	}
-
-	// change the cwd for test
-	err = os.Chdir(tempDir)
-	if err != nil {
-		t.Fatalf("Could not change the temp dir: %v", err)
-	}
-
-	resultFiles, err := GetEnvFiles()
-	if err != nil {
-		t.Fatalf("Error while running GetEnvFiles(): %v", err)
-	}
-	expectedFiles := []string{"global.env", "test.env", "spec_pm.env"}
-	if len(resultFiles) != len(expectedFiles) {
-		t.Errorf(
-			"Expected %d files, got %d files",
-			len(expectedFiles),
-			len(resultFiles),
-		)
-	}
-	/*
-		order is not guranteed with resultFiles so need to create map
-		and check if the file exists
-	*/
-	expectedFilesMap := make(map[string]bool)
-	for _, file := range expectedFiles {
-		expectedFilesMap[file] = true
-	}
-
-	for _, file := range resultFiles {
-		if !expectedFilesMap[strings.ToLower(file)] {
-			t.Errorf("Unexpected file %s returned", file)
-		}
-	}
-}
 
 func TestSetEnvironment(t *testing.T) {
 	/* Test cases
@@ -88,4 +10,9 @@ func TestSetEnvironment(t *testing.T) {
 	* It captures only first argument
 	* second argument should be ignored.
 	 */
+	// environmentFiles := []string{
+	// 	"staging.env",
+	// 	"test_GeneSis.env",
+	// 	"proDuction-A7E.env",
+	// }
 }
