+++ pkg/utils/constants.go
@@ -1,11 +1,17 @@
 package utils
 
 // Colors
-
 const (
 	Red        = "\033[31;1m"
 	Green      = "\033[32;1m"
 	Yellow     = "\033[33;1m"
 	Grey       = "\033[90;1m"
 	colorReset = "\033[0m"
+)
+
+// Environment
+const (
+	EnvKey            = "hulakEnv"
+	DefaultEnvVal     = "global"
+	DefaultFileSuffix = ".env"
 )
